add Location dtos validation

+++ src/Models/Dtos/Location/PutLocationDto.cs
@@ -6,7 +6,7 @@ namespace AirsoftBattlefieldManagementSystemAPI.Models.Dtos.Location
     {
         public decimal? Longitude { get; set; }
         public decimal? Latitude { get; set; }
-        public short? Accuracy { get; set; }
+        public decimal? Accuracy { get; set; }
         public short? Bearing { get; set; }
         public DateTime? Time { get; set; }
     }
